userDTO

+++ models/user/user.go
@@ -4,7 +4,7 @@ type User struct {
 	Id          int    `json:"id"`
 	FirstName   string `json:"first_name"`
 	LastName    string `json:"last_name"`
-	Age         int
+	Age         uint8
 	Email       string
 	Description string
 	Address     string
